+++ test/libdrampowertest/lib_test.cc
@@ -107,7 +107,7 @@ int main(int argc, char* argv[])
         //test.getEnergy().act_stdby_energy
         std::cout << "ACT Std Energy" << "\t" << test.getEnergy().act_stdby_energy << endl;
         //total active standby energy for both ranks
-        std::cout << "ACT Std Energy total ranks" << "\t" << memSpec.memArchSpec.nbrOfRanks *
+        std::cout << "ACT Std Energy total ranks" << "\t" << static_cast<double>(memSpec.memArchSpec.nbrOfRanks) *
         test.getEnergy().act_stdby_energy << "\n" ;
         std::cout << "PRE Std Energy" << "\t" << test.getEnergy().pre_stdby_energy << endl;
         std::cout << "Total Energy" << "\t" << test.getEnergy().total_energy << endl;
